% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ATC_to_text.R
\name{map_ATC_to_text}
\alias{map_ATC_to_text}
\title{Helper to create the mapping table and translate the input given to translate_ATC_codes().}
\usage{
map_ATC_to_text(ATC_codes, len = 7, api_key)
}
\arguments{
\item{ATC_codes}{character vector - ATC codes}

\item{len}{integer - the length of the ATC code level (7 is the maximum 1 the minimum)}

\item{api_key}{character - create an account at https://bioportal.bioontology.org/ and it will come with an api key}
}
\value{
data.frame
}
\description{
Helper to create the mapping table and translate the input given to translate_ATC_codes().
}
\details{
Since the level depth specified in translate_ATC_codes() needs to be translated
         to a string length, this helper was written to perform this task in a controled manner.
}
\examples{
\dontrun{
ATC_vect <- c("A02BC02", "C03BA08", "A02BC02", "A02BC02", "A07DA03", "A07DA03",
              "A02BC02", "A02BC02", "A02BC02", "A02BC02", "C10AA05", "C10AA05",
              "C10AA05", "N05BA06", "N05BA06", "N05BA06", "N06AX11", "N06AX11")
# you will need to create an account at https://bioportal.bioontology.org/
# to receive an api_key
ATCapiR:::map_ATC_to_text(ATC_vect, len = 7, api_key = api_key)
}
}
